kubernetes/video_processing_service/requirements.txt
Flask
mysql-connector-python
kafka
